alert-commands

+++ src/main/java/ru/hse/mmstr_project/se/storage/fast_storage/dto/FriendMetaDto.java
@@ -9,7 +9,7 @@ public record FriendMetaDto(
         String name,
         List<String> wayToNotify,
         String phoneNumber,
-        Integer chatId,
+        Long chatId,
         String email) {
 
     public static FriendMetaDto parse(FriendDto friendDto) {
